Split out the recycler module into its own file.
Much tidier.
