+++ Launchpad/Launchpad.csproj
@@ -40,6 +40,7 @@
   <ItemGroup>
     <Compile Include="AssemblyInfo.cs" />
     <Compile Include="ExLaunchPad.cs" />
+    <Compile Include="Recycler.cs" />
     <Compile Include="Resources.cs" />
   </ItemGroup>
   <ItemGroup>
